+++ handler/micro/micro_test.go
@@ -1,4 +1,4 @@
-package rpc
+package micro
 
 import (
 	"bytes"
